feat: post,put,delete 서버와 통신 확인 완료
