+++ src/function/useImage.js
@@ -1,8 +1,10 @@
 import React, { useState } from 'react';
 
-export const useImageSize = () => {
+export const useImage = () => {
+
   const [selectedImage, setSelectedImage] = useState(null); // 클릭된 이미지 정보를 저장할 상태 변수
   const [isImageSize, setImageSize] = useState(50); //이미지 크기 값 저장
+
   const handleImgSize = (e) => {
     const value = parseInt(e.currentTarget.value); //이벤트로 가져온 value
     if (selectedImage) {
@@ -36,14 +38,12 @@ export const useImageSize = () => {
       // insertImageConfig()
     }
 
-    //=================================================수정
     const fileUrl = URL.createObjectURL(files[0]);
     const fileName = files[0].name;
     // const imgTag = `<img src="${fileUrl}" alt="Uploaded Image" />`;
     // const result = elementRef.current.innerHTML += imgTag;
     insertImageConfig(elementRef, fileUrl, fileName)
 
-    //=================================================수정
 
     return files[0]; //항상 배열로 넘겨줘야함
   };
@@ -65,22 +65,17 @@ export const useImageSize = () => {
   // }
 
   //나중에 export
-  // 고유한 아이디 생성 함수
-  const generateUniqueId = () => {
-    return 'img_' + Math.random().toString(36).slice(2, 11); // 랜덤한 문자열을 이용한 고유 아이디 생성
-  };
 
   // 이미지 삽입, 이미지 설정 함수
   const insertImageConfig = (elementRef, imageDataUrl, elementName) => {
 
     const imgElement = document.createElement('img');
-    const imageId = generateUniqueId(); // 고유한 아이디 생성
 
     imgElement.setAttribute('id', elementName); // 이미지에 아이디 설정
     imgElement.src = imageDataUrl;
 
     // 초기 이미지 크기 값 설정
-    imgElement.style.maxWidth = '50%';
+    imgElement.style.maxWidth = '30%';
     imgElement.style.height = 'auto';
 
     // 이미지 클릭 이벤트 핸들러 추가
