--- src/test/EditPage/EditExamRefactory.js
@@ -1,13 +1,0 @@
-import { useRef } from "react";
-import EditorExam from "../../components/EditorExam"
-// import { useDataHandle } from "../..//dataHandle";
-export default function EditExam() {
-  return (
-    <>
-      <h1>Test</h1>
-      <EditorExam></EditorExam>
-      {/* <EditorExam></EditorExam>
-      <EditorExam></EditorExam> */}
-    </>
-  )
-}
